Add base class for test cases to make things DRY
