+++ phpunit.xml.dist
@@ -10,6 +10,7 @@
 	<testsuites>
 		<testsuite>
 			<directory suffix=".php">./tests/phpunit/tests</directory>
+			<exclude>./tests/phpunit/tests/TestCase.php</exclude>
 		</testsuite>
 	</testsuites>
 	<filter>
