+++ tests/phpunit/tests/ProjectLocator.php
@@ -7,14 +7,13 @@
 
 namespace Required\Traduttore\Tests;
 
-use \GP_UnitTestCase;
 use \Required\Traduttore\Project;
 use \Required\Traduttore\ProjectLocator as Locator;
 
 /**
  * Test cases for \Required\Traduttore\ProjectLocator.
  */
-class ProjectLocator extends GP_UnitTestCase {
+class ProjectLocator extends TestCase {
 	/**
 	 * @var \GP_Project
 	 */
