+++ tests/phpunit/tests/Repository/Bitbucket.php
@@ -7,15 +7,15 @@
 
 namespace Required\Traduttore\Tests\Repository;
 
-use \GP_UnitTestCase;
 use Required\Traduttore\Repository\Bitbucket as BitbucketRepository;
 use \Required\Traduttore\Project;
 use Required\Traduttore\Repository;
+use Required\Traduttore\Tests\TestCase;
 
 /**
  * Test cases for \Required\Traduttore\Repository\Bitbucket.
  */
-class Bitbucket extends GP_UnitTestCase {
+class Bitbucket extends TestCase {
 	/** @var Project */
 	protected $project;
