+++ tests/phpunit/tests/RepositoryFactory.php
@@ -7,7 +7,6 @@
 
 namespace Required\Traduttore\Tests;
 
-use \GP_UnitTestCase;
 use Required\Traduttore\Repository\{Bitbucket, GitHub, GitLab};
 use \Required\Traduttore\Project;
 use \Required\Traduttore\RepositoryFactory as Factory;
@@ -15,7 +14,7 @@ use \Required\Traduttore\RepositoryFactory as Factory;
 /**
  * Test cases for \Required\Traduttore\RepositoryFactory.
  */
-class RepositoryFactory extends GP_UnitTestCase {
+class RepositoryFactory extends TestCase {
 	public function test_get_unknown_repository(): void {
 		$factory = new Factory();
 		$project = new Project(
